# core-lending-loan-servicing-sdk/src/main/resources/api-spec/openapi.yml
openapi: 3.0.1
info:
  title: core-lending-loan-servicing
  description: Loan Servicing Core Application
  contact:
    name: Firefly Software Foundation
    email: dev@getfirefly.io
  version: 1.0.0
servers:
  - url: http://core.getfirefly.io/loan-servicing
    description: Development Environment
  - url: /
    description: Local Development Environment
...
paths:
...
  /api/v1/loan-servicing-cases/{caseId}/rate-changes/{rateChangeId}:
    get:
      tags:
        - LoanRateChange
      summary: Get a loan rate change by ID
      operationId: getRateChangeById
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: rateChangeId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/LoanRateChangeDTO'
    put:
      tags:
        - LoanRateChange
      summary: Update a loan rate change
      operationId: updateRateChange
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: rateChangeId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoanRateChangeDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/LoanRateChangeDTO'
    delete:
      tags:
        - LoanRateChange
      summary: Delete a loan rate change record
      operationId: deleteRateChange
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: rateChangeId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
...
  /api/v1/loan-servicing-cases/{caseId}/rate-changes:
    get:
      tags:
        - LoanRateChange
      summary: List/Search rate changes for a servicing case
      operationId: findAllRateChanges
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: pagination.pageNumber
          in: query
          description: Page number (0-based)
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: pagination.pageSize
          in: query
          description: Number of items per page
          required: false
          schema:
            type: integer
            format: int32
            default: 10
        - name: pagination.sortBy
          in: query
          description: Field to sort by
          required: false
          schema:
            type: string
        - name: pagination.sortDirection
          in: query
          description: Sort direction (ASC or DESC)
          required: false
          schema:
            type: string
            default: DESC
        - name: options.caseInsensitiveStrings
          in: query
          description: Enable case-insensitive string filtering
          required: false
          schema:
            type: boolean
            default: false
        - name: options.includeInheritedFields
          in: query
          description: Include fields from parent classes
          required: false
          schema:
            type: boolean
            default: false
        - name: filters.loanServicingCaseId
          in: query
          description: Exact filter for loan servicing case id
          required: false
          schema:
            type: string
        - name: filters.oldInterestRate
          in: query
          description: Exact filter for old interest rate
          required: false
          schema:
            type: string
        - name: rangeFilters.ranges[oldInterestRate].from
          in: query
          description: Filter old interest rate from value
          required: false
          schema:
            type: string
        - name: rangeFilters.ranges[oldInterestRate].to
          in: query
          description: Filter old interest rate to value
          required: false
          schema:
            type: string
        - name: filters.newInterestRate
          in: query
          description: Exact filter for new interest rate
          required: false
          schema:
            type: string
        - name: rangeFilters.ranges[newInterestRate].from
          in: query
          description: Filter new interest rate from value
          required: false
          schema:
            type: string
        - name: rangeFilters.ranges[newInterestRate].to
          in: query
          description: Filter new interest rate to value
          required: false
          schema:
            type: string
        - name: filters.effectiveDate
          in: query
          description: Exact filter for effective date
          required: false
          schema:
            type: string
        - name: filters.reasonCode
          in: query
          description: Exact filter for reason code
          required: false
          schema:
            type: string
            enum:
              - INDEX_ADJUSTMENT
              - RENEGOTIATION
              - PENALTY
              - PROMOTION
        - name: filters.note
          in: query
          description: Exact filter for note
          required: false
          schema:
            type: string
        - name: filters.createdAt
          in: query
          description: Exact filter for created at
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[createdAt].from
          in: query
          description: Filter created at from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[createdAt].to
          in: query
          description: Filter created at to value
          required: false
          schema:
            type: string
            format: date-time
        - name: filters.updatedAt
          in: query
          description: Exact filter for updated at
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[updatedAt].from
          in: query
          description: Filter updated at from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[updatedAt].to
          in: query
          description: Filter updated at to value
          required: false
          schema:
            type: string
            format: date-time
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PaginationResponseLoanRateChangeDTO'
    post:
      tags:
        - LoanRateChange
      summary: Create a new loan rate change
      operationId: createRateChange
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoanRateChangeDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/LoanRateChangeDTO'
...
components:
  schemas:
...
    LoanRateChangeDTO:
      required:
        - effectiveDate
        - loanServicingCaseId
        - newInterestRate
        - oldInterestRate
        - reasonCode
      type: object
      properties:
        loanRateChangeId:
          type: string
          format: uuid
          readOnly: true
        loanServicingCaseId:
          type: string
          format: uuid
        oldInterestRate:
          maximum: 100
          exclusiveMaximum: false
          minimum: 0
          exclusiveMinimum: false
          type: number
        newInterestRate:
          maximum: 100
          exclusiveMaximum: false
          minimum: 0
          exclusiveMinimum: false
          type: number
        effectiveDate:
          type: string
          format: date
        reasonCode:
          type: string
          enum:
            - INDEX_ADJUSTMENT
            - RENEGOTIATION
            - PENALTY
            - PROMOTION
        note:
          maxLength: 500
          minLength: 0
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    PaginationResponseLoanRateChangeDTO:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            $ref: '#/components/schemas/LoanRateChangeDTO'
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: The current page number, typically zero-based.
          format: int32
      description: Represents a paginated response containing a list of items and pagination metadata.
...